pyproject.toml
[project]
name = "blockchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.128.0",
    "immudb-py>=1.5.0",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "streamlit>=1.52.2",
    "uvicorn>=0.40.0",
]
